Extract VolumeOnly to Cycles namespace
