+++ lib/sof/cycle.rb
@@ -204,30 +204,6 @@ module SOF
 
     def as_json(...) = notation
 
-    class VolumeOnly < self
-      @volume_only = true
-      @notation_id = nil
-      @kind = :volume_only
-      @valid_periods = []
-
-      class << self
-        def handles?(sym) = sym.nil? || super
-
-        def validate_period(period)
-          raise InvalidPeriod, <<~ERR.squish unless period.nil?
-            Invalid period value of '#{period}' provided. Valid periods are:
-            #{valid_periods.join(", ")}
-          ERR
-        end
-      end
-
-      def to_s = "#{volume}x total"
-
-      def covered_dates(dates, ...) = dates
-
-      def cover?(...) = true
-    end
-
     class Lookback < self
       @volume_only = false
       @notation_id = "L"
